improved docs, moved types
